requirements.txt
flask>=3.1.0
openai>=1.72.0
python-dotenv>=1.0.0
gunicorn>=23.0.0
textblob>=0.17.1
ratelimit>=2.2.1
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0
uvicorn>=0.27.0